tweaks

+++ pkg/abstractions/image/image.go
@@ -161,7 +161,7 @@ func (is *RuncImageService) BuildImage(in *pb.BuildImageRequest, stream pb.Image
 	// Use config to override the default clip version (v1 - old format)
 	clipVersion := uint32(clipCommon.ClipFileFormatVersion)
 	if is.config.ImageService.ClipVersion != 0 {
-		clipVersion = uint32(is.config.ImageService.ClipVersion)
+		clipVersion = is.config.ImageService.ClipVersion
 	}
 
 	buildOptions := &BuildOpts{
